requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.1
pluggy==1.5.0
psutil==6.1.0
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.10.0
pymongo==4.10.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-xprocess==1.0.2
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.1
